requirements.txt
python == 3.12.1
python-telegram-bot == 20.8
httpx == 0.26.0
python-dotenv == 1.0.1